Move WS error message to debug log
If the wifi module decides to stop working, the plugin starts
emitting error messages rather constantly, causing the signalk
server to get bogged down and eventually freeze. The normal error
messages are duplicative to the status and the debug messages
anyway, so carry the risk of breaking more things. Just remove
the app.error invocation and move the message to the debug log.